Move boards into component
