+++ components/boards/src/nk3xn.rs
@@ -20,12 +20,9 @@ use lpc55_hal::{
 use memory_regions::MemoryRegions;
 use utils::OptionalStorage;
 
-use crate::{
-    board::Board, flash::ExtFlashStorage, soc::lpc55::Lpc55, store::impl_storage_pointers,
-};
+use crate::{flash::ExtFlashStorage, soc::lpc55::Lpc55, store::impl_storage_pointers, Board};
 
 pub mod button;
-pub mod init;
 pub mod led;
 pub mod nfc;
 pub mod prince;
@@ -42,12 +39,12 @@ use prince::InternalFilesystem;
 use nfc::NfcChip;
 use spi::{FlashCs, Spi};
 
-const MEMORY_REGIONS: &'static MemoryRegions = &MemoryRegions::LPC55;
+pub const MEMORY_REGIONS: &'static MemoryRegions = &MemoryRegions::LPC55;
 
 pub type PwmTimer = ctimer::Ctimer3<Unknown>;
 pub type ButtonsTimer = ctimer::Ctimer1<Unknown>;
 
-type I2C = I2cMaster<
+pub type I2C = I2cMaster<
     Pio0_9,
     Pio1_14,
     I2c5,
@@ -99,46 +96,4 @@ where
         self.0.start(Microseconds::new(delay));
         nb::block!(self.0.wait()).unwrap();
     }
-}
-
-pub fn init(
-    device_peripherals: lpc55_hal::raw::Peripherals,
-    core_peripherals: rtic::export::Peripherals,
-) -> init::All {
-    const SECURE_FIRMWARE_VERSION: u32 = utils::VERSION.encode();
-
-    crate::init_logger::<NK3xN>();
-
-    let hal = lpc55_hal::Peripherals::from((device_peripherals, core_peripherals));
-
-    let require_prince = cfg!(not(feature = "no-encrypted-storage"));
-    let secure_firmware_version = Some(SECURE_FIRMWARE_VERSION);
-    let nfc_enabled = true;
-    let boot_to_bootrom = true;
-
-    init::start(hal.syscon, hal.pmc, hal.anactrl)
-        .next(hal.iocon, hal.gpio)
-        .next(
-            hal.adc,
-            hal.ctimer.0,
-            hal.ctimer.1,
-            hal.ctimer.2,
-            hal.ctimer.3,
-            hal.ctimer.4,
-            hal.pfr,
-            secure_firmware_version,
-            require_prince,
-            boot_to_bootrom,
-        )
-        .next(
-            hal.flexcomm.0,
-            hal.flexcomm.5,
-            hal.inputmux,
-            hal.pint,
-            nfc_enabled,
-        )
-        .next(hal.rng, hal.prince, hal.flash)
-        .next()
-        .next(hal.rtc)
-        .next(hal.usbhs)
 }
