+++ components/boards/src/soc/nrf52/flash.rs
