+++ runners/embedded/Cargo.toml
@@ -11,25 +11,21 @@ license = "MIT"
 [dependencies]
 ### base dependencies
 apps = { path = "../../components/apps" }
+boards = { path = "../../components/boards" }
 cfg-if = "*"
 delog = "0.1"
 cortex-m = { version = "0.7", features = ["critical-section-single-core"]}
 cortex-m-rtic = "1.0"
-embedded-storage = "0.3"
 embedded-hal = "0.2.3"
 embedded-time = "0.12"
-generic-array = "0.14"
-memory-regions = "1"
 nfc-device = { path = "../../components/nfc-device" }
 rand_chacha = { version = "0.3.1", default-features = false, optional = true }
 rand_core = "0.6"
 rtt-target = { version = "0.3", features = ["cortex-m"], optional = true }
-spi-memory = "0.2.0"
 utils = { path = "../../components/utils" }
 serde = { version = "1.0", default-features = false }
 heapless = "0.7"
 heapless-bytes = { version = "0.3.0", features = ["cbor"] }
-lfs-backup = { path = "../../components/lfs-backup" }
 rand = { version =  "0.8.5", default-features = false }
 
 ### protocols and dispatchers
@@ -40,7 +36,6 @@ ctap-types = "0.1"
 ### trussed core
 trussed = "0.1"
 interchange = "0.3"
-littlefs2 = { version = "0.4", features = ["c-stubs"] }
 
 ### usb machinery
 usb-device = "0.2"
@@ -55,7 +50,6 @@ nrf52840-pac = { version = "0.11", optional = true }
 cortex-m-semihosting = { version = "0.3.5", optional = true }
 lpc55-hal = { version = "0.3", features = ["littlefs"], optional = true }
 lpc55-pac = { version = "0.4", optional = true }
-fm11nc08 = { path = "../../components/fm11nc08", optional = true }
 nb = "1"
 systick-monotonic = { version = "1.0.0", optional = true }
 
@@ -76,31 +70,31 @@ default = ["alloc"]
 test = ["apps/test", "utils/test", "se050"]
 develop = ["no-encrypted-storage", "apps/no-reset-time-window", "log-traceP"]
 develop-no-press = ["develop", "no-buttons"]
-provisioner = ["apps/provisioner", "write-undefined-flash", "no-buttons", "apps/no-reset-time-window", "lpc55-hardware-checks"]
+provisioner = ["apps/provisioner", "boards/provisioner", "write-undefined-flash", "no-buttons", "apps/no-reset-time-window", "lpc55-hardware-checks"]
 
 no-delog = ["delog/knock-it-off"]
 
 # Do not use encryption for the filesystem
-no-encrypted-storage = []
+no-encrypted-storage = ["boards/no-encrypted-storage"]
 
 # Check for undefined flash and write to determined value (for prince provisioning)
 write-undefined-flash = []
 
 # Use to auto-succeed every user presence check
-no-buttons = []
+no-buttons = ["boards/no-buttons"]
 
 # Format filesystem anyway
 format-filesystem = []
 
 alloc = ["alloc-cortex-m"]
 
-board-nk3am = ["soc-nrf52"]
-board-nk3xn = ["soc-lpc55"]
+board-nk3am = ["boards/board-nk3am", "soc-nrf52"]
+board-nk3xn = ["boards/board-nk3xn", "soc-lpc55"]
 
 soc-nrf52 = ["nrf52840-hal", "nrf52840-pac", "rand_chacha", "se05x?/nrf"]
-soc-lpc55 = ["lpc55-hal", "lpc55-pac", "fm11nc08", "systick-monotonic", "se05x?/lpc55", "rand_chacha"]
+soc-lpc55 = ["lpc55-hal", "lpc55-pac", "systick-monotonic", "se05x?/lpc55", "rand_chacha"]
 
-se050 = ["dep:se05x", "apps/se050"]
+se050 = ["dep:se05x", "apps/se050", "boards/se050"]
 
 lpc55-hardware-checks = []
